Se crea, actualiza, y sus respectivas traducciones con tags y keywords en el idioma deseado

+++ app/Http/Controllers/PostController.php
@@ -9,6 +9,7 @@ use App\Traits\CategoryPostTrait;
 use App\Traits\ContentTypeTrait;
 use App\Traits\TranslateTrait;
 use App\Traits\LanguageTrait;
+use App\Traits\TaggingTagTrait;
 use App\Traits\PostTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -19,7 +20,7 @@ use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
-  use KeywordTrait; use CategoryPostTrait; use ImageTrait; use ContentTypeTrait; use TranslateTrait; use LanguageTrait;  use PostTrait;
+  use KeywordTrait; use CategoryPostTrait; use TaggingTagTrait; use ImageTrait; use ContentTypeTrait; use TranslateTrait; use LanguageTrait;  use PostTrait;
   public function __construct()
   {
       $this->middleware('auth');
@@ -70,7 +71,33 @@ class PostController extends Controller
                    ->orderBy('created_at','asc')
                    ->get();
                  }
+      $id_lang=$this->getLangIdBySigla(app()->getLocale());
       for($i=0;$i<count($posts);$i++){
+        $tags_lang=[];
+        $keywords_lang=[];
+        $tags=[];
+        $keywords=[];
+        $id_lang_post=$this->getLangIdBySigla($posts[$i]->default_lang);
+        foreach ($posts[$i]->taggingtags as $tag) {
+          if($tag->pivot->language_id===$id_lang){
+            $tags_lang[]=$tag;
+          }
+          if($tag->pivot->language_id===$id_lang_post){
+            $tags[]=$tag;
+          }
+        }
+        for($n=0;$n<count($posts[$i]->keywords);$n++) {
+          if($posts[$i]->keywords[$n]->language_id===$id_lang){
+            $keywords_lang[]=$posts[$i]->keywords[$n];
+          }
+          if($posts[$i]->keywords[$n]->language_id===$id_lang_post){
+            $keywords[]=$posts[$i]->keywords[$n];
+          }
+        }
+        $posts[$i]->tags_lang=$tags_lang;
+        $posts[$i]->tags_post=$tags;
+        $posts[$i]->keywords_lang=$keywords_lang;
+        $posts[$i]->keywords_post=$keywords;
         if($posts[$i]->publicate_state===0){
           $posts[$i]->show=false;
         }
@@ -122,10 +149,9 @@ class PostController extends Controller
         return $categories;
     }
 
-    public function getTags()
+    public function getTagsExisting()
     {
-      $post=Post::first();
-      $tags=$post->existingTags();
+      $tags=$this->getTags();
       $tags_array=[];
       foreach ($tags as $tag) {
         $tags_array[]=array('key'=>'',
@@ -197,22 +223,41 @@ class PostController extends Controller
         $post->cant_shares=0;
         $post->tags=request('tags');
         $post->slug=Str::slug($post->title, '-');
-        $post->keywords=request('keywords');
         $this->authorize('store',$post);
         $post->save();
         $post->qr_img_url='qrcode_'.$post->id.'_'.$post->slug.'.svg';
         $post->update();
         QrCode::format('svg')->color(33, 56, 175)->generate(url('/post-list/'.$post->id),public_path('storage/qrcodes/posts/qrcode_'.$post->id.'_'.$post->slug.'.svg'));
         $post->tag($tags);
+        $lang_id=$this->getLangIdBySigla(request('default-lang'));
+        /*
+         Getting ids from tags post
+        */
+        $ids_tags=[];
+        if(is_array($tags)){
+        foreach ($tags as $tagg) {
+          $tagged=$this->getTagByName($tagg);
+          $ids_tags[]=$tagged[0]->id;
+          $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>$lang_id));
+        }
+      }
+      else{
+        $tagged=$this->getTagByName($tags);
+        $ids_tags[]=$tagged[0]->id;
+        $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>$lang_id));
+      }
+        /*
+         End Getting ids from tags post
+        */
         if(is_array($keywords)){
           foreach($keywords as $keyword){
-            $existKey=$this->getKeywordIf($keyword);
+            $existKey=$this->getKeywordIf($keyword,$lang_id);
 
             $post->keywords()->attach($existKey->id);
           }
         }
         else{
-          $existKey=$this->getKeywordIf($keywords);
+          $existKey=$this->getKeywordIf($keywords,$lang_id);
           $post->keywords()->attach($existKey->id);
         }
 
@@ -230,7 +275,57 @@ class PostController extends Controller
         'checkEditContent'=> 'required',
       ]);
 
+      $tags;
+      $keywords;
+      if(strpos(request('tags'), ',')){
+        $tags = explode(",", request('tags'));
+      }
+      else{
+        $tags=request('tags');
+      }
+
+      if(strpos(request('keywords'), ',')){
+        $keywords = explode(",", request('keywords'));
+      }
+      else{
+        $keywords=request('keywords');
+      }
+
       $post=Post::find(request('post_id'));
+      $post->tag($tags);
+      /*
+       Getting ids from tags post
+      */
+      $ids_tags=[];
+      if(is_array($tags)){
+      foreach ($tags as $tagg) {
+        $tagged=$this->getTagByName($tagg);
+        $ids_tags[]=$tagged[0]->id;
+        $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>request('lang')));
+      }
+    }
+    else{
+      $tagged=$this->getTagByName($tags);
+      $ids_tags[]=$tagged[0]->id;
+      $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>request('lang')));
+    }
+      /*
+       End Getting ids from tags post
+      */
+
+      if(is_array($keywords)){
+        foreach($keywords as $keyword){
+          $existKey=$this->getKeywordIf($keyword,request('lang'));
+
+          $post->keywords()->attach($existKey->id);
+        }
+      }
+      else{
+        $existKey=$this->getKeywordIf($keywords,request('lang'));
+        $post->keywords()->attach($existKey->id);
+      }
+
+
       $this->authorize('storeTranslate',$post);
       $contentType='Post';
       $tipo_content=$this->findContentId($contentType);
@@ -324,7 +419,6 @@ class PostController extends Controller
           'summary'=> 'required',
           'content'=> 'required',
           'tags'=> 'required',
-          'keywords'=> 'required',
         ]);
         $newFileName=$this->manageImage(request('img_url'));
 
@@ -339,14 +433,33 @@ class PostController extends Controller
           'summary'=> 'required',
           'content'=> 'required',
           'tags'=> 'required',
-          'keywords'=> 'required',
         ]);
       }
-
+      $tags;
+      $keywords;
+      if(strpos(request('tags'), ',')){
         $tags = explode(",", request('tags'));
+      }
+      else{
+        $tags=request('tags');
+      }
+
+      if(strpos(request('keywords'), ',')){
         $keywords = explode(",", request('keywords'));
-        $slug = Str::slug($post->title, '-');
-        $post->title=request('title');
+      }
+      else{
+        $keywords=request('keywords');
+      }
+
+        if($post->title!=request('title')){
+        $file_path = public_path("storage/qrcodes/posts/$post->qr_img_url");
+
+       if(File::exists($file_path)) File::delete($file_path);
+       $post->title=request('title');
+       $post->slug=Str::slug($post->title, '-');
+       QrCode::format('svg')->color(33, 56, 175)->generate(url('/post-list/'.$post->id),public_path('storage/qrcodes/posts/qrcode_'.$post->id.'_'.$post->slug.'.svg'));
+       }
+
         $post->default_lang=request('default-lang');
         $post->content=request('content');
         $post->summary=request('summary');
@@ -354,19 +467,63 @@ class PostController extends Controller
         $post->cant_access_read=0;
         $post->cant_likes=0;
         $post->cant_shares=0;
-        $post->tags=request('tags');
-        $post->keywords=request('keywords');
-        $post->slug=Str::slug($post->title, '-');
         $post->qr_img_url='qrcode_'.$post->id.'_'.$post->slug.'.svg';
+        $id_lang_post=$this->getLangIdBySigla($post->default_lang);
+        $tags_lang=[];
+        $tags_id_langs=[];
+        foreach ($post->taggingtags as $tagg) {
+          if($tagg->pivot->language_id!=$id_lang_post){
+            $tags_lang[]=$tagg->name;
+            $tags_id_langs[]=$tagg;
+          }
+        }
+        $tags_lang_text=implode(',',$tags_lang);
+        $new_tags_text=request('tags').','.$tags_lang_text;
+        $post->tags=$new_tags_text;
+        $tags_new=explode(",", $new_tags_text);
         $post->update();
-        QrCode::format('svg')->color(33, 56, 175)->generate(url('/post-list/'.$post->id),public_path('storage/qrcodes/posts/qrcode_'.$post->id.'_'.$post->slug.'.svg'));
-        $post->retag($tags);
+        $post->retag($tags_new);
+        $post->taggingtags()->detach();
+  if(is_array($tags)){
+        foreach ($tags as $tagg) {
+          $tagged=$this->getTagByName($tagg);
+          $ids_tags[]=$tagged[0]->id;
+          $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>$id_lang_post));
+        }
+      }
+      else{
+        $tagged=$this->getTagByName($tags);
+        $ids_tags[]=$tagged[0]->id;
+        $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>$id_lang_post));
+    }
+    foreach ($tags_id_langs as $til) {
+      $tagged=$this->getTagByName($til->name);
+      $ids_tags[]=$til->id;
+      $post->taggingtags()->attach($til->pivot->tag_id,array('language_id'=>$til->pivot->language_id));
+    }
+      $keywords_lang=[];
+      $keywords_id_lang=[];
+      foreach ($post->keywords as $keys) {
+        if($keys->language_id!=$id_lang_post){
+          $keywords_lang[]=$keys->name;
+          $keywords_id_lang[]=$keys;
+        }
+      }
         $post->keywords()->detach();
-        foreach($keywords as $keyword){
-          $existKey=$this->getKeywordIf($keyword);
+        if(is_array($keywords)){
+          foreach($keywords as $keyword){
+            $existKey=$this->getKeywordIf($keyword,$id_lang_post);
 
-
+            $post->keywords()->attach($existKey->id);
+          }
+        }
+        else{
+          $existKey=$this->getKeywordIf($keywords,$id_lang_post);
           $post->keywords()->attach($existKey->id);
+        }
+
+        foreach ($keywords_id_lang as $kil) {
+          $post->keywords()->attach($kil->id);
         }
 
         $postToUpd=$this->getPost($post->id);
@@ -374,17 +531,93 @@ class PostController extends Controller
     }
 
     public function updateTranslatedPostByLang($post_id,$lang_name, Request $request){
-      $dataPost=request()->validate([
+    $dataPost=request()->validate([
         'title'=> 'required|max:255',
         'content'=> 'required',
         'summary'=> 'required',
       ]);
+      $tags;
+      $keywords;
+      if(strpos(request('tags'), ',')){
+        $tags = explode(",", request('tags'));
+      }
+      else{
+        $tags=request('tags');
+      }
 
-      $post=Post::find(request('post_id'));
-      $contentType='Post';
+      if(strpos(request('keywords'), ',')){
+        $keywords = explode(",", request('keywords'));
+      }
+      else{
+        $keywords=request('keywords');
+      }
+
+
+      $post_find=Post::findOrFail($post_id);
+      $post=$this->getPost($post_id);
+        $contentType='Post';
       $tipo_content=$this->findContentId($contentType);
 
       $lang=$this->getLangIdByName($lang_name);
+
+      /*Start keywords and tags process*/
+      $tags_lang=[];
+      $tags_id_langs=[];
+      foreach ($post->taggingtags as $tagg) {
+        if($tagg->pivot->language_id!=$lang){
+          $tags_lang[]=$tagg->name;
+          $tags_id_langs[]=$tagg;
+        }
+      }
+      $tags_lang_text=implode(',',$tags_lang);
+      $new_tags_text=request('tags').','.$tags_lang_text;
+      $post_find->tags=$new_tags_text;
+      $tags_new=explode(",", $new_tags_text);
+      $post_find->update();
+      $post_find->retag($tags_new);
+      $post->taggingtags()->detach();
+if(is_array($tags)){
+      foreach ($tags as $tagg) {
+        $tagged=$this->getTagByName($tagg);
+        $ids_tags[]=$tagged[0]->id;
+        $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>$lang));
+      }
+    }
+    else{
+      $tagged=$this->getTagByName($tags);
+      $ids_tags[]=$tagged[0]->id;
+      $post->taggingtags()->attach($tagged[0]->id,array('language_id'=>$lang));
+  }
+  foreach ($tags_id_langs as $til) {
+    $tagged=$this->getTagByName($til->name);
+    $ids_tags[]=$til->id;
+    $post->taggingtags()->attach($til->pivot->tag_id,array('language_id'=>$til->pivot->language_id));
+  }
+    $keywords_lang=[];
+    $keywords_id_lang=[];
+    foreach ($post->keywords as $keys) {
+      if($keys->language_id!=$lang){
+        $keywords_lang[]=$keys->name;
+        $keywords_id_lang[]=$keys;
+      }
+    }
+      $post->keywords()->detach();
+      if(is_array($keywords)){
+        foreach($keywords as $keyword){
+          $existKey=$this->getKeywordIf($keyword,$lang);
+
+          $post->keywords()->attach($existKey->id);
+        }
+      }
+      else{
+        $existKey=$this->getKeywordIf($keywords,$lang);
+        $post->keywords()->attach($existKey->id);
+      }
+
+      foreach ($keywords_id_lang as $kil) {
+        $post->keywords()->attach($kil->id);
+      }
+      /*End Start keywords y tags process*/
 
 
       $data_trans=array(
